+++ CHANGELOG.md
@@ -7,6 +7,11 @@
   - Pending: Limited Google Drive interoperability
   - Pending: Distinct Server and Client modes
 
+## 0.4.14
+  - Added user current avatar performance to Instance UI
+  - Updated WorldBalancer request paths
+  - Fixed deserialization error during authentication (temporary)
+
 ## 0.4.13
   - Bump dependency `vrchatapi`: 1.20.1 -> 1.20.4
   - Bump dependency `jda`: 5.6.1 -> 6.1.0